2.8 Checking that values which are invertible in DB are NOT NULL.
Name - Cannot be NULL.

Breed - Can be NULL, so there is no need to check that value in the provider.

Gender - Cannot be NULL. Must equal one of these constants: GENDER_MALE, GENDER_FEMALE, or GENDER_UNKNOWN.

Weight - Technically, the weight could be null because of the way we defined our pets table. We added a database constraint to use default value 0 if no weight was provided. So we can allow a null weight value. BUT if a weight value is provided, then we must make sure it’s greater or equal to 0. No negative weights allowed.